ping/pingpong/requirements.txt
Django>=4.2,<5
channels
daphne
djangorestframework
django-cors-headers
channels_redis
